move code to utils
